Add team link to header component and update team routes
